requirements.txt
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
blinker==1.4
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configparser==3.5.0
dispatcher==1.0
enum34==1.1.6
flake8==3.4.1
Flask==0.12.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
pathlib==1.0.1
pew==1.0.2
pipenv==8.2.7
pycodestyle==2.3.1
pyflakes==1.5.0
python-telegram-bot==8.0
requests==2.18.4
requirements==0.1
shutilwhich==1.1.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
Werkzeug==0.12.2
